pyproject.toml
[project]
authors = [{name = "T145", email = "T145@protonmail.com"}]
dependencies = ["ruff>=0.7.1", "asyncclick>=8.1.7.2", "httpx>=0.27.2", "langchain-community>=0.3.4", "langchain-ollama>=0.2.0", "asyncio>=3.4.3", "motor>=3.5.1", "python-dotenv>=1.0.1"]
description = "An all-in-one local AI assistant!"
name = "aida"
requires-python = "> 3.8, < 3.13"
version = "0.1.0"
license.file = "LICENSE"

[project.readme]
file = "README.md"
content-type = "text/markdown"

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.pixi.project]
channels = ["conda-forge", "pytorch", "nvidia", "huggingface", "xformers"]
platforms = ["win-64"]

[tool.pixi.pypi-dependencies]
aida = { path = ".", editable = true }

[tool.pixi.tasks]
start = "python -c 'from aida import main; main.hello()'"

[tool.pixi.dependencies]
pytorch = ">=2.5.1"
torchvision = ">=0.20.1"
torchaudio = ">=2.5.1"
pytorch-cuda = "12.4.*"
langchain = ">=0.3.6"
langgraph = ">=0.2.41"
pymongo = ">=4.10.1"
rich = ">=13.9.3"

[tool.black]
line-length = 132

[tool.mypy]
show_error_context = true
show_column_numbers = true
ignore_missing_imports = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
warn_unused_ignores = true
warn_redundant_casts = true

[tool.ruff]
ignore = ["E501", "E402"]
select = ["E", "F", "W", "I", "B"]
